==> 04_Figures/Figure3/02-R_CD4_Na_Cell_Type_Identification.ipynb <==
quiet_library <- function(...) {
    suppressPackageStartupMessages(library(...))
}
quiet_library(Seurat)
quiet_library(ggplot2)
quiet_library(Matrix)
quiet_library(H5weaver)
quiet_library(dplyr)
quiet_library(viridis)
quiet_library(harmony)
quiet_library(Nebulosa)
quiet_library(ArchR)

cd4_na <- readRDS(file = 'cd4_naive_processed.rds')

# Trim low number clusters

table(cd4_na$`wsnn_res.0.3`)

options(repr.plot.width = 6, repr.plot.height = 6)
DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'wsnn_res.0.3', shuffle = TRUE, label = TRUE)

options(repr.plot.width = 18, repr.plot.height = 15)
p1 <- DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'wsnn_res.0.3', label = TRUE, label.size = 6)
p2 <- DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'pediatric_senior', shuffle = TRUE)
p3 <- DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'age_cmv', shuffle = TRUE)
p4 <- FeaturePlot(cd4_na, reduction = 'wnn.3.umap', features = 'percent.mt')
p1 + p2 + p3 + p4

# DEG Analysis

cd4_na <- SetIdent(cd4_na, value = 'wsnn_res.0.3')
rna_markers <- FindAllMarkers(cd4_na, assay = 'SCT', max.cells.per.ident = 500)

rna_markers %>%
    group_by(cluster) %>%
    top_n(n = 10, wt = avg_log2FC) -> top10

options(repr.plot.width = 18, repr.plot.height = 15)
cd4_na_subset <- subset(cd4_na, downsample = 500)
DoHeatmap(cd4_na_subset, features = top10$gene, assay = 'SCT') + NoLegend()

# Cluster / Cell Types

cd4_na <- SetIdent(cd4_na, value = 'wsnn_res.0.3')
cd4_na <- RenameIdents(cd4_na, '0' = 'CD4 Naive-1',
                          '1' = 'CD4 Naive-2', '2' = 'CD4 Naive-3', '3' = 'CD4 Naive-4', 
                          '4' = 'CD4 SCM',
                          '5' = 'CD25neg Treg',
                          '6' = 'CD4 Naive-5',
                          '7' = 'CD4 Naive-6')
cd4_na$wnn_celltype <- Idents(cd4_na)

options(repr.plot.width = 12, repr.plot.height = 10)
DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'wnn_celltype', label = TRUE, label.size = 8, repel = TRUE)

cd4_na <- SetIdent(cd4_na, value = 'wsnn_res.0.3')
cd4_na <- RenameIdents(cd4_na, '0' = 'CD4 Naive',
                          '1' = 'CD4 Naive', '2' = 'CD4 Naive', '3' = 'CD4 Naive', 
                          '4' = 'CD4 SCM',
                          '5' = 'CD25neg Treg',
                          '6' = 'CD4 Naive',
                          '7' = 'CD4 Naive')
cd4_na$pure_celltype <- Idents(cd4_na)

options(repr.plot.width = 12, repr.plot.height = 10)
DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'pure_celltype', label = TRUE, label.size = 8, repel = TRUE)

options(repr.plot.width = 18, repr.plot.height = 15)
p1 <- DimPlot(cd4_na, reduction = 'umap', group.by = 'pure_celltype', label = TRUE)
p2 <- DimPlot(cd4_na, reduction = 'adtumap', group.by = 'pure_celltype', label = TRUE)
p3 <- DimPlot(cd4_na, reduction = 'atac_umap', group.by = 'pure_celltype', label = TRUE)
p4 <- DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'pure_celltype', label = TRUE)
p1 + p2 + p3 + p4

# Save 

saveRDS(cd4_na, file = 'cd4_na_labeled.rds')

sessionInfo()

# cd4_na <- readRDS(file = '/home/jupyter/Organized_Gating_Analysis/Seurat_Objects/cd4_na.rds')

# saveRDS(cd4_na@meta.data, file = '/home/jupyter/Organized_Gating_Analysis/03_Subset_Deep_Dives/cd4_naive/metadata.rds')
