==> 04_Figures/Figure3/04-R_CD4_Na_Plots.ipynb <==
quiet_library <- function(...) {
    suppressPackageStartupMessages(library(...))
}
quiet_library(Seurat)
quiet_library(ggplot2)
quiet_library(Matrix)
quiet_library(H5weaver)
quiet_library(dplyr)
quiet_library(viridis)
quiet_library(harmony)
quiet_library(Nebulosa)
quiet_library(ArchR)

cd4_na <- readRDS(file = 'cd4_na_labeled.rds')

# CD4 Naive Gate Plot

cd4 <- readRDS(file = '../../Seurat_Objects/cd4_subset.rds')

cd4.batch <- SplitObject(cd4, split.by = 'batch_id')

cd4.batch <- lapply(cd4.batch, function(x){
    NormalizeData(x, assay = 'ADT', normalization.method = 'CLR', margin = 2)})

b069_gating <- cd4.batch$`B069`
nontreg_b069 <- subset(b069_gating, subset = adt_CD25 > 0.8 & adt_CD127 < 1.5, invert = T)

options(repr.plot.width = 8, repr.plot.height = 8)
FeatureScatter(object = nontreg_b069, feature1 = 'adt_CD27', feature2 = 'adt_CD45RA', cols = 'white', pt.size = 0) & geom_density_2d() &
geom_vline(xintercept = 2.75) & geom_hline(yintercept = 3.1) & ggtitle('B069')

# Open a pdf file
pdf("plots/cd4na_gating.pdf", width = 8, height = 8) 
# 2. Create a plot
FeatureScatter(object = nontreg_b069, feature1 = 'adt_CD27', feature2 = 'adt_CD45RA', cols = 'white', pt.size = 0) & geom_density_2d() &
geom_vline(xintercept = 2.75) & geom_hline(yintercept = 3.1) & ggtitle('B069')
# Close the pdf file
dev.off() 

# 3WNN

table(cd4_na$pure_celltype)

options(repr.plot.width = 12, repr.plot.height = 10)
DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'pure_celltype', label = FALSE, label.size = 8, repel = TRUE, cols = c('#006837','#3690c0','#980043'))

# Open a pdf file
pdf("plots/cd4na_celltypes_3WNN_umap.pdf", width = 12, height = 10) 
# 2. Create a plot
DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'pure_celltype', label = FALSE, label.size = 8, repel = TRUE, cols = c('#006837','#3690c0','#980043'))
# Close the pdf file
dev.off() 

options(repr.plot.width = 10, repr.plot.height = 10)
DimPlot(cd4_na, reduction = 'wnn.3.umap', group.by = 'pediatric_senior', label = FALSE, label.size = 8, repel = TRUE, cols = c('#1b9e77','#d95f02'), shuffle = TRUE) & NoLegend()

table(cd4_na$pediatric_senior)

cd4_na <- SetIdent(cd4_na, value = 'pediatric_senior')
downsample_cd4_na <- subset(cd4_na, downsample = 20000)

umap_coords <- as.data.frame(downsample_cd4_na@reductions$wnn.3.umap@cell.embeddings)

head(umap_coords)

umap_df <- data.frame(umap1 = umap_coords$wnn3umap_1,
                      umap2 = umap_coords$wnn3umap_2,
                      age = downsample_cd4_na$pediatric_senior,
                      age2 = paste0(downsample_cd4_na$pediatric_senior,'_2'))

head(umap_df)

options(repr.plot.width = 10, repr.plot.height = 10)
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-7.5,8) +
        ylim(-8,7) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')

# Open a pdf file
pdf("plots/cd4na_celltypes_3WNN_umap_age.pdf", width = 10, height = 10) 
# 2. Create a plot
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-7.5,8) +
        ylim(-8,7) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')
# Close the pdf file
dev.off() 

### ATAC UMAP

umap_coords <- as.data.frame(downsample_cd4_na@reductions$atac_umap@cell.embeddings)

head(umap_coords)

umap_df <- data.frame(umap1 = umap_coords$atacumap_1,
                      umap2 = umap_coords$atacumap_2,
                      age = downsample_cd4_na$pediatric_senior,
                      age2 = paste0(downsample_cd4_na$pediatric_senior,'_2'))

head(umap_df)

options(repr.plot.width = 10, repr.plot.height = 10)
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-6.5,7) +
        ylim(-7,6) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')

# Open a pdf file
pdf("plots/cd4na_celltypes_atac_umap_age.pdf", width = 10, height = 10) 
# 2. Create a plot
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-6.5,7) +
        ylim(-7,6) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')
# Close the pdf file
dev.off() 

### RNA UMAP

umap_coords <- as.data.frame(downsample_cd4_na@reductions$umap@cell.embeddings)

head(umap_coords)

umap_df <- data.frame(umap1 = umap_coords$UMAP_1,
                      umap2 = umap_coords$UMAP_2,
                      age = downsample_cd4_na$pediatric_senior,
                      age2 = paste0(downsample_cd4_na$pediatric_senior,'_2'))

head(umap_df)

options(repr.plot.width = 10, repr.plot.height = 10)
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-7.5,8) +
        ylim(-9.5,7) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')

# Open a pdf file
pdf("plots/cd4na_celltypes_rna_umap_age.pdf", width = 10, height = 10) 
# 2. Create a plot
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-7.5,8) +
        ylim(-9.5,7) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')
# Close the pdf file
dev.off() 

### ADT UMAP

umap_coords <- as.data.frame(downsample_cd4_na@reductions$adtumap@cell.embeddings)

head(umap_coords)

umap_df <- data.frame(umap1 = umap_coords$adtumap_1,
                      umap2 = umap_coords$adtumap_2,
                      age = downsample_cd4_na$pediatric_senior,
                      age2 = paste0(downsample_cd4_na$pediatric_senior,'_2'))

head(umap_df)

options(repr.plot.width = 10, repr.plot.height = 10)
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-5,5) +
        ylim(-5,6) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')

# Open a pdf file
pdf("plots/cd4na_celltypes_adt_umap_age.pdf", width = 10, height = 10) 
# 2. Create a plot
ggplot(umap_df, aes(umap1, umap2)) + 
    geom_density_2d(aes(color = age)) +
    stat_density_2d(aes(alpha = ..level.., fill=age), bins = 6, geom='polygon') +
        scale_fill_manual(values = c('#1b9e77','#d95f02')) + 
        scale_color_manual(values = c('#006837','#d73027')) +
        xlim(-5,5) +
        ylim(-5,6) +
        theme_bw() + 
        theme(axis.text.x = element_blank(),
              axis.text.y = element_blank(),
              axis.title.y = element_blank(),
              axis.title.x = element_blank(),
              legend.position = 'none')
# Close the pdf file
dev.off() 

# Heatmap

sp_merge <- readRDS(file = '../../Seurat_Objects/sp_merge_gating_celltype_updated.rds')

sp_merge <- SetIdent(sp_merge, value = 'gating_celltype')
cd4_nonnaive <- subset(sp_merge, idents = c('CD4 CM','CD4 EM1','CD4 EM2','CD4 TEMRA','Treg'))
cd4_nonnaive$pure_celltype <- cd4_nonnaive$gating_celltype

cd4_merge <- merge(cd4_na, cd4_nonnaive)

cd4_merge <- NormalizeData(cd4_merge, assay = 'RNA')
cd4_merge <- NormalizeData(cd4_merge, assay = 'ADT', margin = 2, normalization.method = 'CLR')
cd4_merge <- SetIdent(cd4_merge, value = 'pure_celltype')

marker_genes_rna <- c('CCR7','CD27','FAS','FOXP3','IL2RA','IFNG')
marker_avg_rna <- AverageExpression(cd4_merge, features = marker_genes_rna, assay = 'RNA')
marker_genes_adt <- c('CD45RA','CD197','CD27','CD49d','CD25')
marker_avg_adt <- AverageExpression(cd4_merge, features = marker_genes_adt, assay = 'ADT')

library(pheatmap)

marker_avg_rna <- marker_avg_rna$RNA
# marker_avg_rna

marker_avg_adt <- marker_avg_adt$ADT
# marker_avg_adt

rownames(marker_avg_adt) <- paste0('adt_',rownames(marker_avg_adt))

# marker_avg_rna
# marker_avg_adt

markers_bind <- rbind(marker_avg_rna,marker_avg_adt)
markers_bind <- as.data.frame(markers_bind)
markers_bind

markers_bind2 <- markers_bind[,c(1,3,4,5,6,7,8,2)]

options(repr.plot.width = 10, repr.plot.height = 8)
cd4na_markers_heatmap <- pheatmap(markers_bind2, scale = 'row', fontsize = 20, cluster_cols = FALSE)

options(repr.plot.width = 10, repr.plot.height = 8)
cd4na_markers_heatmap <- pheatmap(markers_bind2, scale = 'row', fontsize = 20, cluster_cols = TRUE)

save_pheatmap_pdf <- function(x, filename, width=10, height=8) {
   stopifnot(!missing(x))
   stopifnot(!missing(filename))
   pdf(filename, width=width, height=height)
   grid::grid.newpage()
   grid::grid.draw(x$gtable)
   dev.off()
}
save_pheatmap_pdf(cd4na_markers_heatmap, "plots/heatmap.pdf")

## UMAP Plots

library(cowplot)

options(repr.plot.width = 8, repr.plot.height = 8)
# p1 <- FeaturePlot(cd4_na, features = 'cleanadt_CD27', reduction = 'wnn.3.umap', order = TRUE) & scale_color_viridis(limits=c(1,4.5))
# p2 <- FeaturePlot(cd4_na, features = 'cleanadt_CD197', reduction = 'wnn.3.umap', order = TRUE) & scale_color_viridis(limits=c(1,1.75))
# p3 <- FeaturePlot(cd4_na, features = 'cleanadt_CD45RA', reduction = 'wnn.3.umap', order = TRUE) & scale_color_viridis(limits=c(2,5))
p4 <- FeaturePlot(cd4_na, features = 'cleanadt_TCR-Valpha7.2', reduction = 'wnn.3.umap', order = TRUE) & scale_color_viridis()
p5 <- FeaturePlot(cd4_na, features = 'cleanadt_CD49d', reduction = 'wnn.3.umap', order = TRUE) & scale_color_viridis()
# FeaturePlot(cd4_na, features = 'sct_FAS', reduction = 'wnn.3.umap', order = TRUE, pt.size = 1) & scale_color_viridis()

p1 <- plot_density(cd4_na, features = 'sct_SOX4', reduction = 'wnn.3.umap', pal = 'inferno')
p2 <- plot_density(cd4_na, features = 'sct_TOX', reduction = 'wnn.3.umap', pal = 'inferno')
p3 <- plot_density(cd4_na, features = 'sct_DACH1', reduction = 'wnn.3.umap', pal = 'inferno')
p4 <- plot_density(cd4_na, features = 'sct_CPQ', reduction = 'wnn.3.umap', pal = 'inferno')
p5 <- plot_density(cd4_na, features = 'sct_STAT4', reduction = 'wnn.3.umap', pal = 'inferno')
p6 <- plot_density(cd4_na, features = 'sct_INPP4B', reduction = 'wnn.3.umap', pal = 'inferno')

options(repr.plot.width = 20, repr.plot.height = 14)
plot_grid(p1, p2, p3, p4, p5, p6, ncol = 3)

# Open a pdf file
pdf("plots/cd4na_age_markers_umaps.pdf", width = 20, height = 14) 
# 2. Create a plot
plot_grid(p1, p2, p3, p4, p5, p6, ncol = 3)
# Close the pdf file
dev.off() 

p1 <- plot_density(cd4_na, features = 'sct_CACHD1', reduction = 'wnn.3.umap', pal = 'inferno')
p2 <- plot_density(cd4_na, features = 'sct_TGFBR2', reduction = 'wnn.3.umap', pal = 'inferno')
p3 <- plot_density(cd4_na, features = 'sct_PIK3CA', reduction = 'wnn.3.umap', pal = 'inferno')
p4 <- plot_density(cd4_na, features = 'sct_INPP4B', reduction = 'wnn.3.umap', pal = 'inferno')
p5 <- plot_density(cd4_na, features = 'sct_DPP4', reduction = 'wnn.3.umap', pal = 'inferno')
p6 <- plot_density(cd4_na, features = 'sct_GBP5', reduction = 'wnn.3.umap', pal = 'inferno')

options(repr.plot.width = 20, repr.plot.height = 14)
plot_grid(p1, p2, p3, p4, p5, p6, ncol = 3)

# Tracks

Check ATAC notebook 03-R_CD4_Na_ATAC.ipynb

# Subset Counts

split_cd4_na <- SplitObject(cd4_na, split.by = 'pediatric_senior')

celltype_df_func <- function(seurat_object, metadata_slot, id_list){
    seurat_object <- SetIdent(seurat_object, value = metadata_slot)
    subset_object <- subset(seurat_object, idents = id_list[[1]])
    df <- as.data.frame(table(subset_object$pbmc_sample_id))
    colnames(df) <- c('sample',id_list[[1]])
    i = 2
    while(i < length(id_list) + 1){
        subset_object <- subset(seurat_object, idents = id_list[[i]])
        df2 <- as.data.frame(table(subset_object$pbmc_sample_id))
        df[,ncol(df) + 1] <- df2$Freq
        colnames(df)[ncol(df)] <- paste0(id_list[[i]])
        #df$id_list[[i]] <- df2$Freq
        i = i + 1
        }
    df
    }

ped_celltype <- celltype_df_func(split_cd4_na$Pediatric, 'pure_celltype', sort(unique(split_cd4_na$Pediatric$pure_celltype)))

head(ped_celltype)

rownames(ped_celltype) <- ped_celltype$sample
ped_celltype <- ped_celltype[,2:4]
colnames(ped_celltype) <- sort(unique(split_cd4_na$Pediatric$pure_celltype))
ped_celltype

sen_celltype <- celltype_df_func(split_cd4_na$Senior, 'pure_celltype', sort(unique(split_cd4_na$Senior$pure_celltype)))

head(sen_celltype)

rownames(sen_celltype) <- sen_celltype$sample
sen_celltype <- sen_celltype[,2:4]
colnames(sen_celltype) <- sort(unique(split_cd4_na$Senior$pure_celltype))
sen_celltype

write.csv(ped_celltype, file = 'cd4na_freq_ped.csv')
write.csv(sen_celltype, file = 'cd4na_freq_OA.csv')

# DEGs by Age

cd4_na <- NormalizeData(cd4_na, assay = 'RNA')

split_cd4_na <- SplitObject(cd4_na, split.by = 'pure_celltype')

deg_list <- lapply(split_cd4_na, function(x){
    DefaultAssay(x) <- 'RNA'
    x <- SetIdent(x, value = 'pediatric_senior')
    degs <- FindMarkers(x, `ident.1` = 'Pediatric', `ident.2` = 'Senior', min.pct = 0)
    return(degs)
}) 

write.csv(deg_list$`CD4 Naive`, file='cd4na_degs_age.csv')
write.csv(deg_list$`CD4 SCM`, file='cd4scm_degs_age.csv')
write.csv(deg_list$`CD25neg Treg`, file='cd25negtreg_degs_age.csv')

deg_na <- deg_list$`CD4 Naive`
deg_scm <- deg_list$`CD4 SCM`
deg_treg <- deg_list$`CD25neg Treg`

head(deg_na)

deg_na$gene_dir <- paste0(rownames(deg_na),sign(deg_na$avg_log2FC))
head(deg_na)

deg_scm$gene_dir <- paste0(rownames(deg_scm),sign(deg_scm$avg_log2FC))
deg_treg$gene_dir <- paste0(rownames(deg_treg),sign(deg_treg$avg_log2FC))

library(UpSetR)

listInput <- list(Naive = deg_na$gene_dir, SCM = deg_scm$gene_dir, Treg = deg_treg$gene_dir)

options(repr.plot.width = 10, repr.plot.height = 5)
upset(fromList(listInput), order.by = "freq", text.scale = 2)

# Open a pdf file
pdf("plots/deg_upset_plot.pdf", width = 10, height = 5) 
# 2. Create a plot
upset(fromList(listInput), order.by = "freq", text.scale = 2)
# Close the pdf file
dev.off() 

cd4_na <- SetIdent(cd4_na, value = 'pediatric_senior')
downsample_cd4_na <- subset(cd4_na, downsample = 15000)

table(downsample_cd4_na$pediatric_senior)

top10 <- top_n(x=deg_na, n=20, wt=avg_log2FC)

bot10 <- top_n(x=deg_na, n=-23, wt=avg_log2FC)
bot10 <- subset(bot10, !(gene_dir %in% c('AC105402.3-1','MTRNR2L8-1','EDA-1')))
bot10

options(repr.plot.width = 10, repr.plot.height = 10)
DefaultAssay(downsample_cd4_na) <- 'SCT'
DoHeatmap(downsample_cd4_na, features = c(rownames(top10),rownames(bot10)), group.colors = c('#1b9e77','#d95f02')) + 
    scale_fill_gradientn(colors = c('#2b83ba', "black", 'orange'))#d7191c

# Open a pdf file
pdf("plots/deg_heatmap_plot.pdf", width = 10, height = 10) 
# 2. Create a plot
DoHeatmap(downsample_cd4_na, features = c(rownames(top10),rownames(bot10)), group.colors = c('#1b9e77','#d95f02')) + 
    scale_fill_gradientn(colors = c('#2b83ba', "black", 'orange'))
# Close the pdf file
dev.off() 

sort(intersect(intersect(rownames(deg_list$`CD4 Naive`),rownames(deg_list$`CD4 SCM`)),rownames(deg_list$`CD25neg Treg`)))

length(rownames(subset(deg_list$`CD4 Naive`, avg_log2FC > 0)))
length(rownames(subset(deg_list$`CD4 Naive`, avg_log2FC < 0)))

head(subset(deg_list$`CD4 Naive`, avg_log2FC > 0),n=25)
head(subset(deg_list$`CD4 Naive`, avg_log2FC < 0),n=25)

c('TOX','SOX4','IKZF2','CPQ','STAT4','TSHZ2') %in% rownames(deg_list$`CD4 Naive`)

degs_cd4na <- deg_list$`CD4 Naive`
degs_cd4na$genes <- rownames(degs_cd4na)
subset(degs_cd4na, genes %in% c('TOX','SOX4','IKZF2','CPQ','STAT4','TSHZ2'))

length(rownames(subset(deg_list$`CD25neg Treg`, avg_log2FC > 0)))
length(rownames(subset(deg_list$`CD25neg Treg`, avg_log2FC < 0)))

length(rownames(subset(deg_list$`CD4 SCM`, avg_log2FC > 0)))
length(rownames(subset(deg_list$`CD4 SCM`, avg_log2FC < 0)))

df <- data.frame('CD4 Naive' = c(189,-151),
                 'CD4 SCM' = c(186,-207),
                 'CD25neg Treg' = c(175,-206), 
                 age = c('Pediatric','Senior'))
df

tall_df <- tidyr::gather(data = df, key = comparison, value = degs, `CD4.Naive`:`CD25neg.Treg`, factor_key = TRUE)
tall_df$contrast <- 'age'
head(tall_df)

options(repr.plot.width = 8, repr.plot.height = 6)
ggplot(tall_df) + 
    geom_bar(aes(x=comparison, y=degs, fill=age), stat="identity", width=0.5,
             color="black")

library(VennDiagram)

naive_degs <- deg_list$`CD4 Naive`
scm_degs <- deg_list$`CD4 SCM`
treg_degs <- deg_list$`CD25neg Treg`

naive_gene_names <- paste0(sign(naive_degs$avg_log2FC),(rownames(naive_degs)))
scm_gene_names <- paste0(sign(scm_degs$avg_log2FC),(rownames(scm_degs)))
treg_gene_names <- paste0(sign(treg_degs$avg_log2FC),(rownames(treg_degs)))

library(RColorBrewer)
myCol <- brewer.pal(3, "Pastel2")

venn.diagram(
    x = list(naive_gene_names, scm_gene_names, treg_gene_names),
    category.names = c('CD4 Naive', 'CD4 SCM','CD25neg Treg'),
    filename='test.png',
    output=TRUE,
        # Output features
        imagetype="png" ,
        height = 480 , 
        width = 480 , 
        resolution = 300,
        compression = "lzw",
        
        # Circles
        lwd = 2,
        lty = 'blank',
        fill = myCol,
        
        # Numbers
        cex = .6,
        fontface = "bold",
        fontfamily = "sans",
        
        # Set names
        cat.cex = 0.6,
        cat.fontface = "bold",
        cat.default.pos = "outer",
        cat.pos = c(-27, 27, 135),
        cat.dist = c(0.055, 0.055, 0.085),
        cat.fontfamily = "sans",
        rotation = 1
)

# Pseudo-bulk Expression

genes_of_interest <- c('TOX','PABPC1','DACH1','STAT4','INPP4B','CPQ','SESN3','TSHZ2')

cd4_na <- SetIdent(cd4_na, value = 'pure_celltype')
naive_only <- subset(cd4_na, idents = 'CD4 Naive')

split_cd4na <- SplitObject(naive_only, split.by = 'pediatric_senior')

ped_avg_exp <- AverageExpression(object = split_cd4na$Pediatric, assays = 'RNA', features = genes_of_interest, group.by = 'pbmc_sample_id')

sen_avg_exp <- AverageExpression(object = split_cd4na$Senior, assays = 'RNA', features = genes_of_interest, group.by = 'pbmc_sample_id')

write.csv(ped_avg_exp, file = 'cd4na_ped_avg_exp.csv')
write.csv(sen_avg_exp, file = 'cd4na_sen_avg_exp.csv')

# Top 20 Naive DEGs by Age

length(rownames(subset(deg_list$`CD4 Naive`, avg_log2FC > 0)))
length(rownames(subset(deg_list$`CD4 Naive`, avg_log2FC < 0)))

# Split Global UMAP

sp_merge <- readRDS(file = '../../Seurat_Objects/sp_merge_gating_celltype_updated.rds')

sp_merge$pure_celltype <- sp_merge$gating_celltype

## Add CD8na subsets

cd8na_meta <- cd4_na@meta.data

meta_subset_cd8na <- sp_merge@meta.data[which(rownames(sp_merge@meta.data) %in% rownames(cd8na_meta)),]
length(rownames(meta_subset_cd8na))
table(rownames(meta_subset_cd8na) %in% rownames(cd8na_meta))
table(rownames(meta_subset_cd8na) == rownames(cd8na_meta))

Again, it's in the same order, so we don't need to reorder the metadata

sp_merge$pure_celltype[which(rownames(sp_merge@meta.data) %in% rownames(cd8na_meta))] <- as.character(cd8na_meta$pure_celltype)

## Plot

options(repr.plot.width = 25, repr.plot.height = 12)
DimPlot(sp_merge, reduction = 'wnn.3.umap', group.by = 'pure_celltype', raster = FALSE, shuffle = TRUE, split.by = 'pediatric_senior', pt.size = 0.1,
        order = c('CD8 MAIT','CD8 MNP-1','CD8 Naive','CD8 MNP-2','CD8 SCM',
                  'CD4 CM','CD4 EM1','CD4 EM2','CD4 Naive',
                  'CD4 TEMRA','CD4 Unk','CD8 CM','CD8 EM1','CD8 EM2',
                  'CD8 TEMRA','CD8 Unk','Treg'),
        cols = c('lightgrey','lightgrey','lightgrey',
                 'lightgrey','lightgrey','lightgrey','lightgrey','lightgrey',
                 'lightgrey','lightgrey','lightgrey','lightgrey',
                 '#1d91c0','#08519c','#807dba','#01665e','#3f007d'))

# Open a pdf file
pdf("plots/cd8na_celltypes_global_umap.pdf", width = 25, height = 12) 
# 2. Create a plot
DimPlot(sp_merge, reduction = 'wnn.3.umap', group.by = 'pure_celltype', raster = FALSE, shuffle = TRUE, split.by = 'pediatric_senior', pt.size = 0.1,
        order = c('CD8 MAIT','CD8 MNP-1','CD8 Naive','CD8 MNP-2','CD8 SCM',
                  'CD4 CM','CD4 EM1','CD4 EM2','CD4 Naive',
                  'CD4 TEMRA','CD4 Unk','CD8 CM','CD8 EM1','CD8 EM2',
                  'CD8 TEMRA','CD8 Unk','Treg'),
        cols = c('lightgrey','lightgrey','lightgrey',
                 'lightgrey','lightgrey','lightgrey','lightgrey','lightgrey',
                 'lightgrey','lightgrey','lightgrey','lightgrey',
                 '#1d91c0','#08519c','#807dba','#01665e','#3f007d'))
# Close the pdf file
dev.off() 

# Gene Expression DotPlot

gene_list_updated <- c('ITGA6','CD55','ITGB1','ITGA4','KLRD1','ITGAL','KLRC2','KLRC3','IL21R','KLRB1','IL12RB2','KLRG1','LEF1','BACH2','FOXP1','RUNX2','BCL2',
                       'TOX','IKZF2','RUNX3','JUNB','RORA','IFNG-AS1','GZMK','GZMA','PRF1','NCAM1','PECAM1')

cd4_na <- SetIdent(cd4_na, value = 'pure_celltype')
levels(cd4_na) <- c('CD8 MAIT','CD8 SCM','CD8 MNP-2','CD8 MNP-1','CD8 Naive')
levels(cd4_na)

options(repr.plot.width = 12, repr.plot.height = 4)
DotPlot(cd4_na, features = gene_list_updated, 
        col.min = 0, assay = 'SCT') & theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1, size = 20),
                                            axis.text.y = element_text(size = 20),
                                            axis.title.x = element_blank(), axis.title.y = element_blank())

# Open a pdf file
pdf("plots/cd8na_celltypes_RNA_dotplot.pdf", width = 12, height = 4) 
# 2. Create a plot
DotPlot(cd4_na, features = gene_list_updated, 
        col.min = 0, assay = 'SCT') & theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1, size = 20),
                                            axis.text.y = element_text(size = 20),
                                            axis.title.x = element_blank(), axis.title.y = element_blank())
# Close the pdf file
dev.off() 

'#c7e9c0','#74c476','#238b45','#00441b','#6baed6','#02818a','#08519c','#6a51a3','#bdbdbd','#737373'

'#bcbddc','#807dba','#6a51a3','#54278f','#bdbdbd','#08519c','#969696','#ae017e','#3f007d'
